requirements.txt
robotframework==3.2.1
splinter==0.13.0
robotframework-debuglibrary==2.2.0
